+++ sdk/metricsadvisor/ai-metrics-advisor/samples/typescript/src/incidentsAndAlerts.ts
@@ -136,10 +136,10 @@ async function getRootCauses(
 async function listAlerts(client: MetricsAdvisorClient, alertConfigId: string) {
   console.log(`Listing alerts for alert configuration '${alertConfigId}'`);
   console.log("  using for-await-of syntax");
-  for await (const alert of client.listAlertsForAlertConfiguration(
+  for await (const alert of client.listAlerts(
     alertConfigId,
-    new Date("10/22/2020"),
-    new Date("10/24/2020"),
+    new Date("11/01/2020"),
+    new Date("11/05/2020"),
     "AnomalyTime"
   )) {
     console.log("    Alert");
@@ -150,10 +150,10 @@ async function listAlerts(client: MetricsAdvisorClient, alertConfigId: string) {
 
   console.log(`  by pages`);
   const iterator = client
-    .listAlertsForAlertConfiguration(
+    .listAlerts(
       alertConfigId,
-      new Date("10/22/2020"),
-      new Date("10/24/2020"),
+      new Date("11/01/2020"),
+      new Date("11/05/2020"),
       "AnomalyTime"
     )
     .byPage({ maxPageSize: 20 });