+++ sdk/metricsadvisor/ai-metrics-advisor/test/datafeed.spec.ts
@@ -16,7 +16,8 @@ import {
   DataFeedDimension,
   DataFeedMetric,
   MetricsAdvisorAdministrationClient,
-  MetricsAdvisorKeyCredential
+  MetricsAdvisorKeyCredential,
+  UnknownDataFeedSource
 } from "../src";
 import { createRecordedAdminClient, testEnv } from "./util/recordedClients";
 import { Recorder } from "@azure/test-utils-recorder";
@@ -761,6 +762,51 @@ describe("MetricsAdvisorAdministrationClient datafeed", () => {
     it("deletes PostgreSQL data feed", async function() {
       await verifyDataFeedDeletion(client, createdPostGreSqlId);
     });
+
+    it("creates Unknown data feed", async () => {
+      const expectedSource : UnknownDataFeedSource = {
+        dataSourceType: "Unknown",
+        dataSourceParameter: {
+          connectionString: "https://connect-to-postgresql",
+          query: "{ find: postgresql,filter: { Time: @StartTime },batch: 200 }"
+        }
+      };
+      try{
+         await client.createDataFeed({
+          name: postgreSqlFeedName,
+          source: expectedSource,
+          granularity,
+          schema: dataFeedSchema,
+          ingestionSettings: dataFeedIngestion,
+          options
+        });
+        assert.fail("Test should throw error");
+      }
+      catch(error){
+        assert.equal((error as any).message, "Cannot create a data feed with the Unknown source type.");
+      }
+    });
+
+    it("updates data feed to have an unknown data source type", async function() {
+      const patch: DataFeedPatch = {
+        source: {
+          dataSourceType: "Unknown",
+          dataSourceParameter: {
+            connectionString: "https://connect-to-mongodb-patch",
+            database: "data-feed-mongodb-patch",
+            command: "{ find: mongodb,filter: { Time: @StartTime },batch: 200 }"
+          }
+        }
+      };
+      try{
+        await client.updateDataFeed(createdPostGreSqlId, patch);
+        assert.fail("Test should throw error");
+      }
+      catch(error){
+        assert.equal((error as any).message, "Cannot update a data feed to have the Unknown source type.");
+      }
+    });
+    
   });
 }).timeout(60000);